Added OBS submit request support
Use my old OBS C# API to issue OBS REST queries for submit requests
on the project. Report pending requests in a table above linked
packages.

If a request description has a bugzilla ID, replace that with
a link to the bug in the appropriate bugzilla system (defaults
to BNC).